add  auth + images

+++ controllers/recette.js
@@ -1,9 +1,22 @@
 const Recette = require("../models/Recette");
+const fs = require("fs");
 
 exports.createRecette = (req, res, next) => {
-  console.log(req.body);
+  //console.log(req.body.thing);
+  const thingObject = JSON.parse(req.body.recette);
+  delete thingObject._id;
+  delete thingObject._userId;
+
+  console.log(
+    `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+  );
+
   const thing = new Recette({
-    ...req.body,
+    ...thingObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
   thing
     .save()
@@ -24,13 +37,50 @@ exports.getRecette = (req, res, next) => {
 };
 
 exports.updateRecette = (req, res, next) => {
-  Recette.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Recette modifiée !" }))
+  const thingObject = req.file
+    ? {
+        ...JSON.parse(req.body.recette),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : {
+        ...req.body,
+      };
+
+  delete thingObject._userId;
+  Recette.findOne({ _id: req.params.id })
+    .then((recette) => {
+      console.log(recette.author, "-", req.auth.userId);
+      if (recette.author != req.auth.userId) {
+        res.status(401).json({ message: "Non-autorisé" });
+      } else {
+        Recette.updateOne(
+          { _id: req.params.id },
+          { ...thingObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ message: "Recette modifié!" }))
+          .catch((error) => res.status(400).json({ error }));
+      }
+    })
     .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteRecette = (req, res, next) => {
-  Recette.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Recette supprimée !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Recette.findOne({ _id: req.params.id })
+    .then((recette) => {
+      if (recette.author != req.auth.userId) {
+        res.status(401).json({ message: "Non-autorisé" });
+      } else {
+        const filename = recette.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Recette.deleteOne({ _id: req.params.id })
+            .then(() =>
+              res.status(200).json({ message: "Recette supprimée !" })
+            )
+            .catch((error) => res.status(400).json({ error }));
+        });
+      }
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
